requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
pytest
httpx
hypercorn
py-jwt